[HLTriggerOffline] Move Higgs sources in to plugins dir
