+++ HLTriggerOffline/Higgs/plugins/SealModules.cc
@@ -2,6 +2,6 @@
 #include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 
-#include "HLTriggerOffline/Higgs/interface/HLTHiggsValidator.h"
+#include "HLTHiggsValidator.h"
 
 DEFINE_FWK_MODULE(HLTHiggsValidator);
